Create separate test file for each module.
