+++ tests/capFirst.test.js
@@ -1,0 +1,8 @@
+import capFirst from '../src/helpers/capFirst';
+
+describe('capFirst function', () => {
+  test('capitalizes a string', () => {
+    expect(capFirst('hello')).toEqual('Hello');
+    expect(capFirst('hello')).not.toEqual('hello');
+  });
+});
